backend/requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
python-dotenv